+++ app/src/main/java/nextstep/payments/component/card/PaymentCard.kt
@@ -1,4 +1,4 @@
-package nextstep.payments.ui
+package nextstep.payments.component.card
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
