+++ app/src/main/java/nextstep/payments/component/topbar/NewCardTopBar.kt
@@ -1,4 +1,4 @@
-package nextstep.payments.ui
+package nextstep.payments.component.topbar
 
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.filled.ArrowBack
